+++ AWS/tests/connectors/test_abstract_aws_connector.py
@@ -2,10 +2,10 @@
 
 from pathlib import Path
 
-from aws_helpers.client import AwsClient
 from sekoia_automation.connector import DefaultConnectorConfiguration
 
 from connectors import AbstractAwsConnector, AwsModule
+from connectors.s3.provider import AwsAccountProvider
 
 
 def test_abstract_aws_connector(aws_module: AwsModule, symphony_storage: Path, intake_key: str):
@@ -18,4 +18,4 @@ def test_abstract_aws_connector(aws_module: AwsModule, symphony_storage: Path, i
     connector = AbstractAwsConnector(module=aws_module, data_path=symphony_storage)
     connector.configuration = DefaultConnectorConfiguration(intake_key=intake_key)
 
-    assert isinstance(connector.aws_client, AwsClient)
+    assert isinstance(connector, AwsAccountProvider)
